# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateMapSummaryRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      record_date:
        type: string
      score_count:
        type: integer
      showcase:
        type: string
      user_name:
        type: string
    required:
    - category_id
    - description
    - record_date
    - score_count
    - user_name
    type: object
  handlers.DeleteMapSummaryRequest:
    properties:
      route_id:
        type: integer
    required:
    - route_id
    type: object
...
  handlers.EditMapImageRequest:
    properties:
      image:
        type: string
    required:
    - image
    type: object
  handlers.EditMapSummaryRequest:
    properties:
      description:
        type: string
      record_date:
        type: string
      route_id:
        type: integer
      score_count:
        type: integer
      showcase:
        type: string
      user_name:
        type: string
    required:
    - description
    - record_date
    - route_id
    - score_count
    - user_name
    type: object
...
  handlers.MapSummaryResponse:
    properties:
      map:
        $ref: '#/definitions/models.Map'
      summary:
        $ref: '#/definitions/models.MapSummary'
    type: object
...
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Map:
    properties:
      chapter_name:
        type: string
      game_name:
        type: string
      id:
        type: integer
      image:
        type: string
      is_coop:
        type: boolean
      is_disabled:
        type: boolean
      map_name:
        type: string
    type: object
  models.MapHistory:
    properties:
      date:
        type: string
      runner_name:
        type: string
      score_count:
        type: integer
    type: object
  models.MapRoute:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      completion_count:
        type: integer
      description:
        type: string
      history:
        $ref: '#/definitions/models.MapHistory'
      rating:
        type: number
      route_id:
        type: integer
      showcase:
        type: string
    type: object
...
  models.MapSummary:
    properties:
      routes:
        items:
          $ref: '#/definitions/models.MapRoute'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: lp.pektezol.dev
info:
  contact: {}
  description: Backend API endpoints for Least Portals Hub.
  license:
    name: GNU Affero General Public License, Version 3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Least Portals Hub
  version: "1.0"
paths:
...
  /maps/{mapid}/image:
    put:
      description: Edit map image with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EditMapImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EditMapImageRequest'
              type: object
      tags:
      - maps / summary
...
  /maps/{mapid}/summary:
    delete:
      description: Delete map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.DeleteMapSummaryRequest'
              type: object
      tags:
      - maps / summary
    get:
      description: Get map summary with specified id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MapSummaryResponse'
              type: object
      tags:
      - maps / summary
    post:
      description: Create map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateMapSummaryRequest'
              type: object
      tags:
      - maps / summary
    put:
      description: Edit map summary with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EditMapSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EditMapSummaryRequest'
              type: object
      tags:
      - maps / summary
...
swagger: "2.0"